requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
PyYAML>=6.0.0

# Data Fetching
yfinance>=0.2.30
requests>=2.31.0

# API & Web Server
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0.0

# Database & Environment
supabase>=2.3.0
python-dotenv>=1.0.0

# Dashboard & Visualization
streamlit>=1.30.0
plotly>=5.18.0

# Optional/Advanced
# pandas_ta>=0.3.14  # Uncommon technical indicators (Currently causing install issues)
scipy>=1.11.0
